+++ concrete_datastore/routes/views.py
@@ -1,7 +1,10 @@
 # coding: utf-8
-from importlib import import_module
 import yaml
 import json
+import base64
+import qrcode
+import qrcode.image.svg
+
 from django.conf import settings
 from django.http import (
     HttpResponse,
@@ -9,8 +12,8 @@ from django.http import (
     HttpResponseForbidden,
     StreamingHttpResponse,
 )
-
 from django.views.generic import TemplateView
+from django.contrib.auth import get_user_model
 
 from rest_framework.views import APIView
 from rest_framework import authentication
@@ -19,8 +22,11 @@ from concrete_datastore.api.v1.authentication import (
     TokenExpiryAuthentication,
     URLTokenExpiryAuthentication,
 )
+
 import concrete_datastore
 from concrete_datastore.interfaces.yaml_renderer import DatamodelYamlToHtml
+from concrete_datastore.concrete.constants import MFA_OTP
+from concrete_datastore.routes.forms import ConfigureOTPLoginForm
 from concrete_datastore.interfaces.openapi_schema_generator import (
     SchemaGenerator,
 )
@@ -147,3 +153,37 @@ class OpenApiView(APIView):
             return HttpResponse(spec.decode())
 
         return JsonResponse(schema)
+
+
+class ConfigureOTPView(TemplateView):
+    template_name = 'mainApp/configure-otp.html'
+    form_class = ConfigureOTPLoginForm
+    success_url = '.'
+
+    def post(self, request, *args, **kwargs):
+        User = get_user_model()
+        context = self.get_context_data()
+        form = context['form']
+        if form.is_valid():
+            email = form.cleaned_data['email']
+            user = User.objects.get(email=email)
+            context['user_is_authenticated'] = True
+            device, _ = user.emaildevice_set.get_or_create(
+                mfa_mode=MFA_OTP, confirmed=True
+            )
+            img = qrcode.make(
+                device.config_url, image_factory=qrcode.image.svg.SvgImage
+            )
+
+            base64_qrcode_img_bytes = base64.b64encode(img.to_string())
+            base64_qrcode_img = base64_qrcode_img_bytes.decode('utf-8')
+            context['base64_qrcode_img'] = base64_qrcode_img
+
+        return super().render_to_response(context)
+
+    def get_context_data(self, **kwargs):
+        context = super(ConfigureOTPView, self).get_context_data(**kwargs)
+        form = ConfigureOTPLoginForm(self.request.POST or None)
+        context["form"] = form
+        context["platform_name"] = settings.PLATFORM_NAME
+        return context